+++ SQLServerCourse.Domain/EntitiesForPracticalTasks/Ticket.cs
@@ -11,9 +11,9 @@ namespace SQLServerCourse.Domain.Entitys_for_lesson
     {
         public long Id { get; set; }
 
-        public byte Row { get; set; }
+        public int Row { get; set; }
 
-        public byte Seat { get; set; }
+        public int Seat { get; set; }
 
         public decimal Cost { get; set; }
